mm-reasoner/requirements.txt
fschat
openai
numpy
azure-ai-vision